Make iceberg-477 compile

+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/catalog/glue/GlueIcebergTableOperations.java
@@ -13,16 +13,6 @@
  */
 package io.trino.plugin.iceberg.catalog.glue;
 
-import com.amazonaws.services.glue.AWSGlueAsync;
-import com.amazonaws.services.glue.model.AlreadyExistsException;
-import com.amazonaws.services.glue.model.ConcurrentModificationException;
-import com.amazonaws.services.glue.model.CreateTableRequest;
-import com.amazonaws.services.glue.model.EntityNotFoundException;
-import com.amazonaws.services.glue.model.InvalidInputException;
-import com.amazonaws.services.glue.model.ResourceNumberLimitExceededException;
-import com.amazonaws.services.glue.model.Table;
-import com.amazonaws.services.glue.model.TableInput;
-import com.amazonaws.services.glue.model.UpdateTableRequest;
 import com.google.common.collect.ImmutableMap;
 import io.trino.plugin.hive.metastore.glue.GlueMetastoreStats;
 import io.trino.plugin.iceberg.UnknownTableTypeException;
@@ -37,18 +27,33 @@ import org.apache.iceberg.TableMetadata;
 import org.apache.iceberg.exceptions.CommitFailedException;
 import org.apache.iceberg.exceptions.CommitStateUnknownException;
 import org.apache.iceberg.io.FileIO;
+import software.amazon.awssdk.services.glue.GlueClient;
+import software.amazon.awssdk.services.glue.model.AlreadyExistsException;
+import software.amazon.awssdk.services.glue.model.ConcurrentModificationException;
+import software.amazon.awssdk.services.glue.model.EntityNotFoundException;
+import software.amazon.awssdk.services.glue.model.GlueException;
+import software.amazon.awssdk.services.glue.model.InvalidInputException;
+import software.amazon.awssdk.services.glue.model.ResourceNumberLimitExceededException;
+import software.amazon.awssdk.services.glue.model.StorageDescriptor;
+import software.amazon.awssdk.services.glue.model.Table;
+import software.amazon.awssdk.services.glue.model.TableInput;
 
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+import java.util.function.BiFunction;
 
 import static com.google.common.base.Verify.verify;
-import static io.trino.plugin.hive.ViewReaderUtil.isHiveOrPrestoView;
-import static io.trino.plugin.hive.ViewReaderUtil.isPrestoView;
-import static io.trino.plugin.hive.metastore.glue.converter.GlueToTrinoConverter.getTableParameters;
-import static io.trino.plugin.hive.metastore.glue.converter.GlueToTrinoConverter.getTableType;
 import static io.trino.plugin.hive.util.HiveUtil.isIcebergTable;
+import static io.trino.plugin.iceberg.IcebergErrorCode.ICEBERG_COMMIT_ERROR;
 import static io.trino.plugin.iceberg.IcebergErrorCode.ICEBERG_INVALID_METADATA;
+import static io.trino.plugin.iceberg.IcebergTableName.isMaterializedViewStorage;
+import static io.trino.plugin.iceberg.IcebergTableName.tableNameFrom;
+import static io.trino.plugin.iceberg.IcebergUtil.isTrinoMaterializedView;
+import static io.trino.plugin.iceberg.IcebergUtil.isTrinoView;
+import static io.trino.plugin.iceberg.catalog.glue.GlueIcebergUtil.getMaterializedViewTableInput;
 import static io.trino.plugin.iceberg.catalog.glue.GlueIcebergUtil.getTableInput;
+import static io.trino.plugin.iceberg.catalog.glue.GlueIcebergUtil.getTableType;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 import static org.apache.iceberg.BaseMetastoreTableOperations.METADATA_LOCATION_PROP;
@@ -59,7 +64,7 @@ public class GlueIcebergTableOperations
 {
     private final TypeManager typeManager;
     private final boolean cacheTableMetadata;
-    private final AWSGlueAsync glueClient;
+    private final GlueClient glueClient;
     private final GlueMetastoreStats stats;
     private final GetGlueTable getGlueTable;
 
@@ -69,7 +74,7 @@ public class GlueIcebergTableOperations
     protected GlueIcebergTableOperations(
             TypeManager typeManager,
             boolean cacheTableMetadata,
-            AWSGlueAsync glueClient,
+            GlueClient glueClient,
             GlueMetastoreStats stats,
             GetGlueTable getGlueTable,
             FileIO fileIo,
@@ -90,15 +95,25 @@ public class GlueIcebergTableOperations
     @Override
     protected String getRefreshedLocation(boolean invalidateCaches)
     {
-        Table table = getTable(invalidateCaches);
-        glueVersionId = table.getVersionId();
+        boolean isMaterializedViewStorageTable = isMaterializedViewStorage(tableName);
 
-        Map<String, String> parameters = getTableParameters(table);
-        if (isPrestoView(parameters) && isHiveOrPrestoView(getTableType(table))) {
-            // this is a Presto Hive view, hence not a table
+        Table table;
+        if (isMaterializedViewStorageTable) {
+            table = getTable(database, tableNameFrom(tableName), invalidateCaches);
+        }
+        else {
+            table = getTable(database, tableName, invalidateCaches);
+        }
+        glueVersionId = table.versionId();
+
+        String tableType = getTableType(table);
+        Map<String, String> parameters = table.parameters();
+        if (!isMaterializedViewStorageTable && (isTrinoView(tableType, parameters) || isTrinoMaterializedView(tableType, parameters))) {
+            // this is a Hive view or Trino/Presto view, or Trino materialized view, hence not a table
+            // TODO table operations should not be constructed for views (remove exception-driven code path)
             throw new TableNotFoundException(getSchemaTableName());
         }
-        if (!isIcebergTable(parameters)) {
+        if (!isMaterializedViewStorageTable && !isIcebergTable(parameters)) {
             throw new UnknownTableTypeException(getSchemaTableName());
         }
 
@@ -114,21 +129,22 @@ public class GlueIcebergTableOperations
     {
         verify(version.isEmpty(), "commitNewTable called on a table which already exists");
         String newMetadataLocation = writeNewMetadata(metadata, 0);
-        TableInput tableInput = getTableInput(typeManager, tableName, owner, metadata, newMetadataLocation, ImmutableMap.of(), cacheTableMetadata);
+        TableInput tableInput = getTableInput(typeManager, tableName, owner, metadata, metadata.location(), newMetadataLocation, ImmutableMap.of(), cacheTableMetadata);
 
-        CreateTableRequest createTableRequest = new CreateTableRequest()
-                .withDatabaseName(database)
-                .withTableInput(tableInput);
         try {
-            stats.getCreateTable().call(() -> glueClient.createTable(createTableRequest));
+            stats.getCreateTable().call(() -> glueClient.createTable(x -> x
+                    .databaseName(database)
+                    .tableInput(tableInput)));
         }
-        catch (AlreadyExistsException
-               | EntityNotFoundException
-               | InvalidInputException
-               | ResourceNumberLimitExceededException e) {
-            // clean up metadata files corresponding to the current transaction
-            fileIo.deleteFile(newMetadataLocation);
-            throw e;
+        catch (GlueException e) {
+            if (e instanceof AlreadyExistsException ||
+                    e instanceof EntityNotFoundException ||
+                    e instanceof InvalidInputException ||
+                    e instanceof ResourceNumberLimitExceededException) {
+                // clean up metadata files corresponding to the current transaction
+                fileIo.deleteFile(newMetadataLocation);
+            }
+            throw new TrinoException(ICEBERG_COMMIT_ERROR, "Cannot commit table creation", e);
         }
         shouldRefresh = true;
     }
@@ -136,22 +152,50 @@ public class GlueIcebergTableOperations
     @Override
     protected void commitToExistingTable(TableMetadata base, TableMetadata metadata)
     {
-        String newMetadataLocation = writeNewMetadata(metadata, version.orElseThrow() + 1);
-        TableInput tableInput = getTableInput(
-                typeManager,
-                tableName,
-                owner,
+        commitTableUpdate(
+                getTable(database, tableName, false),
                 metadata,
-                newMetadataLocation,
-                ImmutableMap.of(PREVIOUS_METADATA_LOCATION_PROP, currentMetadataLocation),
-                cacheTableMetadata);
+                (table, newMetadataLocation) ->
+                        getTableInput(
+                                typeManager,
+                                tableName,
+                                owner,
+                                metadata,
+                                Optional.ofNullable(table.storageDescriptor()).map(StorageDescriptor::location).orElse(null),
+                                newMetadataLocation,
+                                ImmutableMap.of(PREVIOUS_METADATA_LOCATION_PROP, currentMetadataLocation),
+                                cacheTableMetadata));
+    }
 
-        UpdateTableRequest updateTableRequest = new UpdateTableRequest()
-                .withDatabaseName(database)
-                .withTableInput(tableInput)
-                .withVersionId(glueVersionId);
+    @Override
+    protected void commitMaterializedViewRefresh(TableMetadata base, TableMetadata metadata)
+    {
+        commitTableUpdate(
+                getTable(database, tableNameFrom(tableName), false),
+                metadata,
+                (table, newMetadataLocation) -> {
+                    Map<String, String> parameters = new HashMap<>(table.parameters());
+                    parameters.put(METADATA_LOCATION_PROP, newMetadataLocation);
+                    parameters.put(PREVIOUS_METADATA_LOCATION_PROP, currentMetadataLocation);
+
+                    return getMaterializedViewTableInput(
+                            table.name(),
+                            table.viewOriginalText(),
+                            table.owner(),
+                            parameters);
+                });
+    }
+
+    private void commitTableUpdate(Table table, TableMetadata metadata, BiFunction<Table, String, TableInput> tableUpdateFunction)
+    {
+        String newMetadataLocation = writeNewMetadata(metadata, version.orElseThrow() + 1);
+        TableInput tableInput = tableUpdateFunction.apply(table, newMetadataLocation);
+
         try {
-            stats.getUpdateTable().call(() -> glueClient.updateTable(updateTableRequest));
+            stats.getUpdateTable().call(() -> glueClient.updateTable(x -> x
+                    .databaseName(database)
+                    .tableInput(tableInput)
+                    .versionId(glueVersionId)));
         }
         catch (ConcurrentModificationException e) {
             // CommitFailedException is handled as a special case in the Iceberg library. This commit will automatically retry
@@ -159,7 +203,7 @@ public class GlueIcebergTableOperations
         }
         catch (EntityNotFoundException | InvalidInputException | ResourceNumberLimitExceededException e) {
             // Signal a non-retriable commit failure and eventually clean up metadata files corresponding to the current transaction
-            throw e;
+            throw new TrinoException(ICEBERG_COMMIT_ERROR, "Cannot commit table update", e);
         }
         catch (RuntimeException e) {
             // Cannot determine whether the `updateTable` operation was successful,
@@ -169,7 +213,7 @@ public class GlueIcebergTableOperations
         shouldRefresh = true;
     }
 
-    private Table getTable(boolean invalidateCaches)
+    private Table getTable(String database, String tableName, boolean invalidateCaches)
     {
         return getGlueTable.get(new SchemaTableName(database, tableName), invalidateCaches);
     }
